added ObjectId validation

+++ controllers/tasks.js
@@ -1,8 +1,18 @@
 const { Task } = require("../models/task");
-const { ctrlWrapper, handleHttpError } = require("../helpers");
+const {
+  ctrlWrapper,
+  handleHttpError,
+  mongooseObjectIdCheck,
+} = require("../helpers");
 
 const getTaskById = async (req, res) => {
   const { id } = req.params;
+  if (!mongooseObjectIdCheck(id)) {
+    throw handleHttpError(
+      400,
+      `Bad request, id isn't match ObjectId mongoose type`
+    );
+  }
   const result = await Task.findById(id, "-createdAt -updatedAt");
   if (!result) {
     throw handleHttpError(404, `Task with id: ${id} is not found`);
@@ -12,6 +22,18 @@ const getTaskById = async (req, res) => {
 
 const createTask = async (req, res) => {
   const { boardId, columnId } = req.params;
+  if (!mongooseObjectIdCheck(boardId)) {
+    throw handleHttpError(
+      400,
+      `Bad request, boardId isn't match ObjectId mongoose type`
+    );
+  }
+  if (!mongooseObjectIdCheck(columnId)) {
+    throw handleHttpError(
+      400,
+      `Bad request, columnId isn't match ObjectId mongoose type`
+    );
+  }
   const result = await Task.create({
     ...req.body,
     board: boardId,
@@ -22,6 +44,12 @@ const createTask = async (req, res) => {
 
 const updateTask = async (req, res) => {
   const { id } = req.params;
+  if (!mongooseObjectIdCheck(id)) {
+    throw handleHttpError(
+      400,
+      `Bad request, id isn't match ObjectId mongoose type`
+    );
+  }
   const result = await Task.findByIdAndUpdate(id, req.body, {
     new: true,
   });
@@ -33,6 +61,12 @@ const updateTask = async (req, res) => {
 
 const deleteTask = async (req, res) => {
   const { id } = req.params;
+  if (!mongooseObjectIdCheck(id)) {
+    throw handleHttpError(
+      400,
+      `Bad request, id isn't match ObjectId mongoose type`
+    );
+  }
   const result = await Task.findByIdAndDelete(id);
   if (!result) {
     throw handleHttpError(404, `Task with id: ${id} is not found`);
@@ -42,6 +76,12 @@ const deleteTask = async (req, res) => {
 
 const filterTasksByPriority = async (req, res) => {
   const { priority, columnId } = req.params;
+  if (!mongooseObjectIdCheck(columnId)) {
+    throw handleHttpError(
+      400,
+      `Bad request, columnId isn't match ObjectId mongoose type`
+    );
+  }
   const result = await Task.find({ column: columnId, priority });
   res.json(result);
 };
